test(shell): the disposal-race cases join XDebuggerView.test.ts under the handlers they drive
`disposal-handling.test.ts` tested only `XDebuggerView`, under the same
top-level title as `XDebuggerView.test.ts`. Its six cases now live in that
file, one `describe()` per handler under `instance members`, with the
rationale for the disposal-race suppression as the block's comment and the
helpers at module level. The former file's header restated the view's
contract, so nothing of it is kept as a header. Each renamed case has its
alias line in `tasks/test-identity-aliases.jsonl`.
