+++ tasks/test-identity-aliases.jsonl
@@ -876,3 +876,9 @@
 {"date":"2026-09-08","from":{"k":"unit","s":"runner","n":"cell-cache: two-identity shared-space compile cache (e2e) > cold writeback recovers from a partial source cache already committed by another replica"},"to":{"k":"unit","s":"runner","n":"cell-cache > two-identity shared-space compile cache (e2e) > cold writeback recovers from a partial source cache already committed by another replica"}}
 {"date":"2026-09-08","from":{"k":"unit","s":"runner","n":"cell-cache: code documents forged below the commit boundary > treats a compiled record whose code document holds other content as a miss"},"to":{"k":"unit","s":"runner","n":"cell-cache > code documents forged below the commit boundary > treats a compiled record whose code document holds other content as a miss"}}
 {"date":"2026-09-11","from":{"k":"unit","s":"runner","n":"Cell with Optional Link > ensureLink() error handling > should throw error when accessing cell without frame context"},"to":{"k":"unit","s":"runner","n":"Cell with Optional Link > ensureLink() error handling > throws naming the missing space for a cell whose link and frame both lack one"}}
+{"date":"2026-09-12","from":{"k":"unit","s":"shell","n":"XDebuggerView > resetBaseline logs a failure while the runtime is alive"},"to":{"k":"unit","s":"shell","n":"XDebuggerView > instance members > resetBaseline() > logs a failure while the runtime is alive"}}
+{"date":"2026-09-12","from":{"k":"unit","s":"shell","n":"XDebuggerView > resetBaseline stays silent when the runtime is disposed"},"to":{"k":"unit","s":"shell","n":"XDebuggerView > instance members > resetBaseline() > stays silent when the runtime is disposed"}}
+{"date":"2026-09-12","from":{"k":"unit","s":"shell","n":"XDebuggerView > toggleLogger logs a failure while the runtime is alive"},"to":{"k":"unit","s":"shell","n":"XDebuggerView > instance members > toggleLogger() > logs a failure while the runtime is alive"}}
+{"date":"2026-09-12","from":{"k":"unit","s":"shell","n":"XDebuggerView > toggleLogger stays silent when the runtime is disposed"},"to":{"k":"unit","s":"shell","n":"XDebuggerView > instance members > toggleLogger() > stays silent when the runtime is disposed"}}
+{"date":"2026-09-12","from":{"k":"unit","s":"shell","n":"XDebuggerView > setLoggerLevel logs a failure while the runtime is alive"},"to":{"k":"unit","s":"shell","n":"XDebuggerView > instance members > setLoggerLevel() > logs a failure while the runtime is alive"}}
+{"date":"2026-09-12","from":{"k":"unit","s":"shell","n":"XDebuggerView > setLoggerLevel stays silent when the runtime is disposed"},"to":{"k":"unit","s":"shell","n":"XDebuggerView > instance members > setLoggerLevel() > stays silent when the runtime is disposed"}}
